Break down scene specific code into modules
