+++ src/app/features/runtimes/react-1v1/scenes/IdleScene.tsx
@@ -1,0 +1,3 @@
+export function IdleScene() {
+  return null;
+}
